feat: Implement module permissions and enhance role management
- Added ModulePermissionController for managing module permissions.
- Introduced ModulePermissionService for handling business logic related to module permissions.
- Created ModulePermission and Permission resources for API responses.
- Updated RoleController to include permissions retrieval and synchronization.
- Enhanced RoleService to manage role permissions and validation.
- Added new validation requests for syncing permissions and role creation/updating.
- Implemented session check and user retrieval in AuthController.
- Updated routes to include new endpoints for module permissions and role permissions.
- Added migration for module_permissions table and updated permissions table to include module_permission_id.
- Created factories and seeders for module permissions and updated existing seeders for permissions.

+++ app/Http/Controllers/Api/V1/Role/RoleController.php
@@ -7,6 +7,8 @@ use App\Models\Role;
 use App\Http\Requests\Role\RoleListRequest;
 use App\Http\Requests\Role\RoleCreateRequest;
 use App\Http\Requests\Role\RoleUpdateRequest;
+use App\Http\Requests\Role\SyncPermissionRequest;
+use App\Http\Resources\Permission\PermissionResource;
 use App\Services\RoleService;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Http\Resources\Json\JsonResource;
@@ -17,9 +19,8 @@ class RoleController extends Controller
 {
 
     public function __construct(
-        private readonly RoleService $roleService
-    ) {
-        $this->authorizeResource(Role::class, 'role');
+        public readonly RoleService $roleService
+    ) {   
     }
 
     /**
@@ -32,6 +33,9 @@ class RoleController extends Controller
      */
     public function index(RoleListRequest $request): AnonymousResourceCollection
     {
+        //
+        $this->authorize('viewAny', Role::class);
+        
         $roles = $this->roleService->index($request);
 
         return RoleResource::collection($roles);
@@ -47,6 +51,8 @@ class RoleController extends Controller
      */
     public function store(RoleCreateRequest $request): JsonResource
     {
+        $this->authorize('create', Role::class);
+
         $role = $this->roleService->store($request);
 
         return RoleResource::make($role);
@@ -62,6 +68,8 @@ class RoleController extends Controller
      */
     public function show(Role $role): JsonResource
     {
+        $this->authorize('view', $role);
+
         $role = $this->roleService->show($role);
 
         return RoleResource::make($role);
@@ -78,6 +86,8 @@ class RoleController extends Controller
      */
     public function update(RoleUpdateRequest $request, Role $role): JsonResource
     {
+        $this->authorize('update', $role);
+
         $role = $this->roleService->update($request, $role);
 
         return RoleResource::make($role);
@@ -92,9 +102,46 @@ class RoleController extends Controller
      * @return Response
      */
     public function destroy(Role $role): Response
-     {
+    {
+        $this->authorize('delete', $role);
+
         $this->roleService->destroy($role);
 
         return response()->noContent();
-     }
+    }
+
+    /**
+     * Get Role Permissions
+     * 
+     * Display the permissions of the specified resource.
+     * 
+     * @param  Role  $role
+     * @return JsonResource
+     */
+    public function permissions(Role $role): JsonResource
+    {
+        $this->authorize('view', $role);
+
+        $permissions = $this->roleService->permissions($role);
+
+        return PermissionResource::collection($permissions);
+    }
+
+    /**
+     * Sync Role Permissions
+     * 
+     * Sync the permissions of the specified resource.
+     * 
+     * @param  SyncPermissionRequest  $request
+     * @param  Role  $role
+     * @return JsonResource
+     */
+    public function syncPermissions(SyncPermissionRequest $request, Role $role): JsonResource
+    {
+        $this->authorize('update', $role);
+
+        $this->roleService->syncPermissions($request,$role);
+
+        return RoleResource::make($role);
+    }
 }
