+++ app/Services/RoleService.php
@@ -6,6 +6,7 @@ use App\Models\Role;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Http\Request;
 use Illuminate\Pagination\AbstractPaginator;
+use Illuminate\Validation\ValidationException;
 
 class RoleService
 {
@@ -17,7 +18,10 @@ class RoleService
      */
     public function index(Request $request): Collection | AbstractPaginator
     {
-        $roles = (new Role())->search($request);
+        $roles = (new Role())->search(
+            request:$request,
+            filters: ['name', 'description'],
+        );
 
         return $roles;
     }
@@ -30,7 +34,11 @@ class RoleService
      */
     public function store(Request $request): Role
     {
-        return Role::create($request->validated());
+        return Role::create([
+            'name' => $request->name,
+            'description' => $request->description,
+            'guard_name' => 'web',
+        ]);
     }
 
     /**
@@ -68,6 +76,38 @@ class RoleService
      */
     public function destroy(Role $role): void
     {
+        // Verify if the role is not assigned to any user before deleting
+        if ($role->users()->count() > 0) {
+            throw ValidationException::withMessages([
+                'role' => [__('role.cannot_delete_assigned')],
+            ]);
+        }
+
         $role->delete();
+    }
+
+    /**
+     * Get permissions of a role.
+     * 
+     * @param  Role  $role
+     * @return Collection
+     */
+    public function permissions(Role $role): Collection
+    {
+        return $role->permissions;
+    }
+
+    /**
+     * Sync permissions for a role.
+     * 
+     * @param  Role  $role
+     * @param  Request  $request
+     * @return void
+     */
+    public function syncPermissions(Request $request, Role $role): void
+    {
+        $permissions = $request->array('permissions.*.id');
+
+        $role->syncPermissions($permissions);
     }
 }
